src/cloud_functions/ingestion/gsheet_ingestion/requirements.txt
functions-framework==3.*
airbyte
google-cloud-storage
google-cloud-secret-manager
